# OnlineShopBackend-main/src/main/resources/static/swagger.yml
openapi: 3.0.3
info:
  title: Online Shop
  description: |-
    Aplicación que se encarga de habilitar endpoints para el consumo REST de la base de datos en la tienda online.
  termsOfService: http://swagger.io/terms/
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
...
servers:
  - url: http://localhost:8080/api/v1
    description: Local server
...
paths:
...
  /shop/create-product:
    post:
      tags:
        - products
      summary: Crear un producto
      description: Crear un producto
      operationId: createProduct
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ProductRequest'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          description: Invalid input
        '422':
          description: Validation exception
        '500':
          description: Internal Server Error - IOException occurred
  /shop/update-product/{productId}:
    patch:
      tags:
        - products
      summary: Actualizar un producto
      description: Actualizar un producto
      operationId: updateProduct
      parameters:
        - name: productId
          in: path
          description: Id del producto
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ProductRequest'
          application/json:
            schema:
              $ref: '#/components/schemas/ProductRequest'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          description: Invalid input
        '422':
          description: Validation exception
        '500':
          description: Internal Server Error - IOException occurred
  /shop/product/{productId}:
    get:
      tags:
        - products
      summary: Encontrar un producto por ID
      description: Econtrar un producto por su ID correspondiente.
      operationId: getProductById
      parameters:
        - name: productId
          in: path
          description: ID del producto
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          description: Invalid ID supplied
        '404':
          description: Product not found
  /shop/product/delete-product/{productId}:
    delete:
      tags:
        - products
      summary: Eliminar un producto por ID
      description: Eliminar un producto por su ID correspondiente.
      operationId: deleteProductById
      parameters:
        - name: productId
          in: path
          description: ID del producto
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
        '400':
          description: Invalid ID supplied
        '404':
          description: Product not found
...
  /shop/carts:
    get:
      tags:
        - cart
      summary: Obtener todos los carritos de compra
      description: Devuelve una lista de todos los carritos de compra.
      operationId: getAllCarts
      responses:
        '200':
          description: Operación exitosa, devuelve una lista de carritos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CartResponse'
        '400':
          description: Solicitud incorrecta
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /shop/cart/{cartId}:
    get:
      tags:
        - cart
      summary: Obtener carrito de la compra
      description: Devuelve el carrito de la compra
      operationId: getCart
      parameters:
        - name: cartId
          in: path
          description: 'Id del carrito a buscar '
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
        '404':
          description: Carrito no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /shop/create-cart:
    post:
      tags:
        - cart
      summary: Crear un nuevo carrito de la compra
      description: Crea un nuevo carrito de la compra y devuelve el carrito creado
      operationId: createCart
      requestBody:
        description: 'Datos para la creación del carrito'
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CartRequest'
      responses:
        '200':
          description: Carrito creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
        '400':
          description: Solicitud incorrecta
  /shop/cart/{cartId}/add-product:
    post:
      tags:
        - cart
      summary: Añadir producto al carrito de la compra
      description: Devuelve el carrito de la compra actualizado
      operationId: addCartProduct
      parameters:
        - name: cartId
          in: path
          description: 'Id del carrito a buscar'
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: 'RequestBody de los detalles del carrito'
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CartDetailsRequest'
      responses:
        '200':
          description: Operación exitosa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartDetailsResponse'
        '404':
          description: El carrito no ha podido ser creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /shop/cart/{cartId}/update-cart:
    put:
      tags:
        - cart
      summary: Actualizar información del carrito de la compra
      description: Devuelve el carrito de la compra actualizado
      operationId: updateCart
      parameters:
        - name: cartId
          in: path
          description: 'Id del carrito a actualizar'
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: 'RequestBody con la información a actualizar del carrito'
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CartRequest'
      responses:
        '200':
          description: Operación exitosa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
        '404':
          description: Carrito no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Solicitud incorrecta
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /shop/cart/delete-cart/{cartId}:
    delete:
      tags:
        - cart
      summary: Eliminar carrito de la compra
      description: Elimina el carrito de la compra identificado por el `cartId`.
      operationId: deleteCart
      parameters:
        - name: cartId
          in: path
          description: 'ID del carrito a eliminar'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Carrito eliminado exitosamente
          content:
            text/plain:
              schema:
                type: string
                example: "El carrito ha sido eliminado con éxito"
        '404':
          description: Carrito no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Solicitud incorrecta
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /shop/cart/{cartId}/remove-product/{cartDetailsId}:
    delete:
      tags:
        - cart
      summary: Eliminar un producto del carrito de la compra
      description: Elimina un producto del carrito de la compra
      operationId: removeCartProduct
      parameters:
        - name: cartId
          in: path
          description: 'Id del carrito a buscar'
          required: true
          schema:
            type: integer
            format: int64
        - name: cartDetailsId
          in: path
          description: 'Id de los detalles del carrito'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Operación exitosa
          content:
            text/plain:
              schema:
                type: string
                example: "El producto ha sido eliminado del carrito con éxito"
...
  /shop/cart/products/{customerId}:
    get:
      tags:
        - cart
      summary: Obtener productos del carrito de un usuario
      description: Devuelve la lista de productos en el carrito de un usuario por su customerId.
      operationId: getProductsInCartByCustomerId
      parameters:
        - name: customerId
          in: path
          description: ID del usuario (customer)
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Lista de productos en el carrito
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductResponse'
        '404':
          description: Carrito no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
...
    CartRequest:
      type: object
      properties:
        cartDetails:
          type: array
          items:
            $ref: '#/components/schemas/CartDetailsRequest'
        deliveryAddress:
          type: string
          example: "Calle Falsa 123, Ciudad"
        paymentMethod:
          type: string
          example: "creditCard"
        date:
          type: string
          format: date-time
          example: "2024-05-26T12:30:00Z"
      required:
        - date
        - deliveryAddress
        - paymentMethod
    CartResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example:
        userId:
          type: integer
          format: int64
          example: 1
        date:
          type: string
          format: date-time
          example: "2024-05-26T12:30:00Z"
        cartDetails:
          type: array
          items:
            $ref: '#/components/schemas/CartDetailsResponse'
    CartDetailsRequest:
      type: object
      properties:
        productId:
          type: integer
          format: int64
          example: 2
        quantity:
          type: integer
          example: 1
    CartDetailsResponse:
      type: object
      properties:
        productId:
          type: integer
          format: int64
          example: 2
        price:
          type: number
          example: 29.99
        quantity:
          type: integer
          example: 1
...
    ProductRequest:
      type: object
      properties:
        name:
          type: string
          example: name
        description:
          type: string
          example: descripcion del producto
        size:
          type: string
          example: "24 x 24"
        type:
          type: string
          example: cuadro
        price:
          type: number
          example: 20.99
        image:
          type: string
          format: binary
          example: photo.jpg
    ProductResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: name
        description:
          type: string
          example: descripcion del producto
        size:
          type: string
          example: "24 x 24"
        type:
          type: string
          example: cuadro
        price:
          type: number
          format: float
          example: 20.99
        image:
          type: string
          example: photo.jpg
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
          example: 404
          description: "Código de estado HTTP que indica el tipo de error."
        message:
          type: string
          example: "Cart not found"
          description: "Mensaje descriptivo del error que proporciona más detalles."
        details:
          type: string
          example: "The cart with the given ID does not exist."
          description: "Información adicional o detalles sobre el error (opcional)."